requirements.txt
python-telegram-bot==20.7
selenium==4.16.0
selenium-stealth==1.0.6
webdriver-manager==4.0.1
flask==3.0.0
gunicorn==21.2.0